fix: Create Project Version 2 for socket streaming
