+++ client/src/components/StudentDashboard.jsx
@@ -5,20 +5,18 @@ import AboutMe from './AboutMe';
 import ExpertSearchComponent from './WhoCanHelp';
 import StudentProjects from './StudentProjects/StudentProjects';
 import SidebarMorningPulse from './MorningPulse';
-import CreateProjectAI from './CreateProject/CreateProjectAI';
-import StreamingProjectCreator from './CreateProject/StreamingProjectCreator';
+import CreateProject from './CreateProject/CreateProject';
 
 export default function StudentDashboard({ email }) {
   return (
     <div>
       <IntroductionContainer />
       <SidebarMorningPulse />
-      <CreateProjectAI />
+      <CreateProject/>
       <StudentProjects />
       <SidebarAdvice />
       <ExpertSearchComponent />
       <AboutMe />
-      <StreamingProjectCreator/>
       <SidebarBreakTimer />
     </div>
   );
